fix(ui): fixed sidebar button tooltip animation bug and general polish (#2080)
There was a bug for the sidebar button tooltips where the framer motion
animation was overwriting the floating UI transforms. There was a fix in
place, but the fix still had an issue the first time the user mouses
into the button. The tooltip would animate across the front of button,
but only the first time the user mouses in.

This fix simplifies it by adding a positional component wrapping the
framer motion animation.

![CleanShot 2025-05-09 at 15 55
13@2x](https://github.com/user-attachments/assets/5e7e142d-74dc-43b8-bd15-cd2060149bc9)

![CleanShot 2025-05-09 at 15 58
49@2x](https://github.com/user-attachments/assets/fa25ddcc-0ba8-4a80-9c6d-43ebaa49fc71)

I also made a few style tweaks...

- Fixed line-height on `h6` tags within tooltips. 
in general, the line-height on h6 using the `<Typography />` component
will cause alignment issues everywhere in the layout.... but I didn't
want to change it globally since im focused on fixing it for the
tooltips. This is the main problem with design system components like
this, if I fix it there, I have no idea where else in the app I might
break something.

- Made the Gem icon larger in the logged out state
- Made the Gift image larger
- Added hover state to the Disconnect button

Here's a video of the Bug, notice how the tooltip animates across the
front of the button on the first run, then works after that:

https://www.loom.com/share/9468680acf0e4f3ba3aad56c18d8af5c?sid=fd411a2c-b8cc-45b2-a043-4096886e2a76

Here's a video of the fix:
https://www.loom.com/share/95a973606f5c46b0a45355cfcea2a7fb




<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Style**
- Increased the height of the gift image in the Invite sidebar for
improved visual appearance.
	- Updated the gem image in the Log In sidebar to a larger version.
- Enhanced tooltip design with new animation, positioning, and styling
for a more polished user experience.
- Added underline effect to buttons on hover within tooltips for clearer
interactivity.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->