feat: Improve swap processing modal
Adds a `noHeader` prop to the TransactionModal to conditionally render the header. This allows for a cleaner UI in cases where a custom header is desired, like in the swap processing modal. Refactors the swap processing modal to use a common status list component, status hero, and a processing icon. Fixes dark mode styling issue. Sets the default value of fromAmount to 1. Disable the wallet action buttons.