refactor: Clear swap details on dialog close
This commit refactors the swap data clearing logic.

Previously, `clearCalculatedDetails` was called in several places within `useSwapData`, including inside the `try...catch` block of `calculateTrade`. This was causing issues when the processing dialog was closed, as it would sometimes clear the details before the transaction was fully completed.

Now, `clearCalculatedDetails` is only called when the `processingOpen` prop changes to `false` in the `ProcessingTransaction` component. This ensures that swap details are cleared only when the swap dialog is closed, preventing premature clearing and race conditions, and gives user better visual confirmation that swap has been completed.