fix(ui): sync page coin video stream from cloudflare (#2277)
This loads the coin video as an adaptive bitrate stream from cloudflare
( HLS.js ) instead of being baked into the app. Loading the video from
the app was causing some issues, it also increased bundle sizes.

This shows a static image while the streaming video is loading then
seamlessly ( pretty close ) switches from static image to video when
it's ready.

Downsides:

- If the user has no internet connection they will only see the static
image ( but I don't think they can sync in this case anyways ? )
- If their connection is slow, the quality of the video will
automatically reduce to adapt to the slow connection, good and bad
thing.

Upsides:

- Better performance when loading the video
- Less overall app bundle size

Note:

I had to slightly change the background color of the app on Dark Mode to
match the video. The difference is so slight it shouldn't make a
difference.

![CleanShot 2025-06-03 at 13 26
18@2x](https://github.com/user-attachments/assets/e79e70e3-3e63-46c2-95f8-0b4ec99e38a2)

---------

Co-authored-by: shan <47271333+shanimal08@users.noreply.github.com>
Co-authored-by: Shannon Tenner <shanimal@tari.com>