feat: Integrate Swap iframe
This commit introduces an iframe to embed the Swap functionality into the wallet. The iframe is hosted on a separate domain. Communication is handled through `postMessage` to facilitate requests and responses between the parent and iframe, including handling wallet connect requests. New styles were introduced for the iframe and iframe container in order to provide a good UX on varying screen sizes, including handling wallet connect requests.