chore: log error to sentry if elevating permissions fails (#2317)
Description
---
Sometimes on macOS users fails to launch `minotari_node` receiving os
error 61: refused to connect to TCP port. This most likely is a result
of `minotari_node` failing to launch which is caused by lack of execute
permissions on the file.
I was not able to recreate it so to get any data regarding underlying
issue we need to increase logging of this issue for further analysis.