fix(ui): wallet ui improvements (#2409)
This PR adds additional polish to the wallet UI. The main updates are a
smooth animation for the shrink scroll interaction and I moved the Sync
with Phone button to be an icon in the corner of the wallet. Other than
those, lots of little adjustments here and there to prevent layout shift
and improve the overall alignment.

Moved the "Sync with Phone" button here:

![CleanShot 2025-06-17 at 18 44
10@2x](https://github.com/user-attachments/assets/7c002e62-969c-43a5-b93d-328206ab9874)

![CleanShot 2025-06-17 at 18 43
30@2x](https://github.com/user-attachments/assets/efc6046e-585b-4dab-87e0-9b7094d15c2c)

Updated Scroll Shrink Animation:

![CleanShot 2025-06-17 at 18 44
58@2x](https://github.com/user-attachments/assets/6253a808-d0a0-416c-acd0-348817ea0904)

Video:


https://www.loom.com/share/0b83632199e243dd82000feda0e77355?sid=1b8e504b-421b-4705-87c6-552a74ea9a3f