chore(ui): [FE TD] Tari address validation (#2451)
Description
---

- just made a reusable hook for tari address (and amount) validation

unfortunately couldn't make a reusable input _component_ because as i
started i realised every single address input 1. is designed differently
and 2. behaves differently in the context of the form it's in ☹️

Motivation and Context
---
- closes #2251 


How Has This Been Tested?
---
- locally

What process can a PR reviewer use to test or verify this change?
---
- check that address (and amount) validation is still working

---------

Co-authored-by: Brian Pearce <brianp@users.noreply.github.com>