fix(ui): hide exchange button on scroll and show my balance text (#2468)
Just a few small tweaks to the Wallet section. 

If the Available Balance is the same as the main balance, it now shows
"My Balance" label. It's weird to show the same value twice if available
balance and mined balance is the same.

![CleanShot 2025-07-02 at 16 13
39@2x](https://github.com/user-attachments/assets/252aa291-2953-40b8-9d71-d0d5db8f15a7)

I also made it so the Exchange button hides itself when the History List
is scrolled, this was a request from design.

![CleanShot 2025-07-02 at 16 15
29@2x](https://github.com/user-attachments/assets/cf1e9131-35f0-4da9-ad8d-d7842df242f7)