fix(ui): move send and receive buttons to the filter items row (#2480)
The Send and Receive buttons seem out of place in the current UI. They
are not part of the latest Figma designs and were carried over from the
previous build before the "Buy Tari" button.

Their size is large and distracting, disrupting the visual flow of the
layout.

I suggest shrinking the buttons and moving them to the empty space next
to the "All Activity" filter dropdown. This approach addresses two
issues at once and creates a clearer appearance. Additionally, usability
improves since the buttons can remain visible when the UI shrinks after
scrolling.

Overall, this change results in a cleaner look and enhances the user
experience.


Before:

![CleanShot 2025-07-03 at 18 08
04@2x](https://github.com/user-attachments/assets/edf925a4-a3fb-4d16-9f61-77b9077824c3)

After:

![CleanShot 2025-07-03 at 18 02
40@2x](https://github.com/user-attachments/assets/2e81a3ec-e6ec-423e-b54c-0760e83f42d8)

![CleanShot 2025-07-03 at 18 28
03@2x](https://github.com/user-attachments/assets/379fb24c-9c8e-49af-a16c-0f43bb7b805b)

Scrolled:

![CleanShot 2025-07-03 at 18 02
57@2x](https://github.com/user-attachments/assets/e8f65388-8215-4bed-af80-5763689ce70a)

---------

Co-authored-by: Shannon Tenner <shanimal@tari.com>