feat(ui): animated gradients on loading elements (#2486)
I added a new `LoadingText` component that adds a subtle animated
gradient to whatever text is passed to it. This makes the text itself
function as a loading indicator. ( This is commonly seen in AI apps like
Claude )

![CleanShot 2025-07-07 at 16 50
04@2x](https://github.com/user-attachments/assets/4eed4e93-ee64-4c11-8b7d-a6d8b9543a96)

I also updated the main Progress bar to have an animated background as
well.

![CleanShot 2025-07-07 at 16 48
21@2x](https://github.com/user-attachments/assets/435f43b1-aa88-4875-abab-fad19c67cd14)

Dark mode 

![CleanShot 2025-07-07 at 16 52
23@2x](https://github.com/user-attachments/assets/7a160f1f-89d9-4b7a-96fe-16ce6760b668)

![CleanShot 2025-07-07 at 16 53
16@2x](https://github.com/user-attachments/assets/9f8ebc76-be0a-4fa2-b2a5-236e28afef3e)