fix: seed word restoration (#2495)
Description
---


Solves the issue that under linux we would have deleted a folder in
`app_local_data_dir` which is different than the app_config_dir()
(.local/share as opposed to .config). The fix does not solve the
underlying issue that we are trying to delete the folder but prevents
the app from crashing if the folder is not there (we do nothing in that
case)

Motivation and Context
---

Fixes #2487 


How Has This Been Tested?
---

Locally withe esme



What process can a PR reviewer use to test or verify this change?
---


Change the seed words while setting the application up.


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify