fix: base node db cleanup (#2502)
Description
---

Fix #2305 the issue where we were trying to delete the peer db folder
from under the wrong folder.
In linux the following code segments:

```
// in linux this is ~/.local/share/com.tari.universe/
   let local_data_dir = app
                .path()
                .app_local_data_dir()
                .expect("Could not get local data dir");
```

And 

```
// in linux this is ~/.config/com.tari.universe/
    let config_path = app
                .path()
                .app_config_dir()
                .expect("Could not get config dir");
```

are two pointing to two different places however in the code this was
used interchangeably which is wrong as we might place something in the
config folder while trying to delete or look it up later on from the
other folder.

Motivation and Context
---

How Has This Been Tested?
---

Locally. 

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify