feat(bridge): extend Tapplet Signer functionality (#2507)
Description
---

- In the new Bridge version [the new function of showing history
](https://github.com/tari-project/wxtm-bridge-backend/issues/102) will
be added. Therefore, the `TappletSigner` was extended with the
`getBackendBridgeTxs` to allow fetch the data from the TU if needed.
- The wallet balance can be taken from the store, so it was refactored
to make sure that the same amount is show on the wallet and in the
Bridge app.
- Also the redundant code has been removed.


Motivation and Context
---
[Add transaction history to the
Brodge](https://github.com/tari-project/wxtm-bridge-backend/issues/102)

How Has This Been Tested?
---
Manually with the 1.2.16 version of the TU.

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

---------

Co-authored-by: Brian Pearce <brianp@users.noreply.github.com>