feat: Add tari address to telemetry data
This commit adds the Tari address to the telemetry data being sent. This will allow the team to better understand which wallets are using the app and how they are interacting with the Tari network. The telemetry_api_url has also been reverted back to the tari.com endpoint to align with the production environment.