fix: refresh balance after initial scan is completed (#2633)
Description
---
Fixes: #2564 

After importing wallet, the scanned balance was invalid but wallet would
re-validate it after around ~30 seconds. To fix it I added flat 30
seconds timeout to check balances again and change it.


How Has This Been Tested?
---
Imported wallet and confirmed that balance was correct.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->