fix(ui): tower freeze post success (#2667)
Description
---

the tower was getting stuck that we've added the success animation state
for pool rewards - this introduced the possibility of having one
"result" state right after the other (after a long block). i.e.
`success` after a pool reward, then `fail` due to not winning the last
block, and vice versa.

- the [tower PR](https://github.com/tari-project/tari-tower/pull/27) has
more context, but the basic fix was just to prevent adding the same
"status" type into the queue if it's already there
- also made a small change for exchange miner logos to use `||` instead
of `??` for fallbacks in case the items aren't completely falsy:D

Motivation and Context
---
- closes https://github.com/tari-project/universe/issues/2643


How Has This Been Tested?
---

- TU locally:

on TU it's kind of tough to get into the state, but this shows the real
`fail` `result`(status) being respected, and not getting into a stuck
state because of the `success` `result`(status):


https://github.com/user-attachments/assets/1c7290e9-5411-43ff-b17f-4518c9944fad

----

- on tower repo locally:

 by spam clicking ALL the `result` status options in quick succession: