feat: Display additional crew members to nudge
This commit adds the functionality to display crew members who need a nudge but are not yet part of the crew list. This is achieved by introducing a `membersToNudge` prop to the `CrewList` component, which is populated from the airdrop store. These members are then rendered in a separate section within the `CrewList` component.