feat(ui): security reminder warning button (#2668)
Instead of auto-triggering the staged security modal multiple times,
let's add this button to the wallet in the sidebar.

The button will no longer be visible once all the steps are completed.

With this new button in the UI, the modal should only auto-pop once. If
a user Skips or closes it, it should not auto trigger again.

<img width="1218" height="696" alt="67005"
src="https://github.com/user-attachments/assets/edf0cd5d-03e4-4979-91fa-a89e123ff4c1"
/>

---------

Co-authored-by: Shannon Tenner <shanimal@tari.com>
Co-authored-by: shan <47271333+shanimal08@users.noreply.github.com>
Co-authored-by: Brian Pearce <brianp@users.noreply.github.com>
Co-authored-by: brianp <brian.o.pearce@gmail.com>