fix(ui): add state animation to the sidebar (#2699)
The sidebar was not animating properly when open/closed. There were some
props there for an animation, but i'm taking a more simplified approach
here.

The sidebar now animates when it opens and exits. The effect is subtle
but adds more visual flare as opposed to popping in and out.

Here's a video of the animation. 


https://www.loom.com/share/c71e99aefcb9412b908e6c7b907ad77c?sid=dae3c0df-901a-4c91-b532-1bef541e5af7

<img width="2678" height="1540" alt="CleanShot 2025-08-11 at 15 23
22@2x"
src="https://github.com/user-attachments/assets/ec66053b-03e9-4d13-9fc0-a157aae4e280"
/>