refactor: Simplify mining status websocket sending
Removed unnecessary `if let` binding for `jwt_token` and `app_version_option` within the `tokio::select!` block. Instead, the `jwt` and `app_version` variables are now initialized *before* the select, using `map_or` to provide default values if the options are `None`. This simplifies the code and makes it easier to read.