fix(ui): add minimize button to crew rewards widget (#2745)
There was no way to minimize or close the Crew Rewards widget, which was
a little annoying, it can get in the way.

This PR adds a `MinimizeButton` that just shrinks it down to a button in
the top corner of the app. This allows the user to "hide" the widget
when they don't need it.

We may want to store this minimized state as a settings variable in the
future. For now it will reset when the app is relaunched.

<img width="1078" height="522" alt="CleanShot 2025-08-18 at 22 15 47@2x"
src="https://github.com/user-attachments/assets/6b47ea5a-58bd-481c-9129-935edd9268f1"
/>

<img width="836" height="526" alt="CleanShot 2025-08-18 at 22 16 13@2x"
src="https://github.com/user-attachments/assets/79d0cbd5-4131-40db-8a1b-43b8e95bb460"
/>

<img width="956" height="782" alt="CleanShot 2025-08-18 at 22 24 46@2x"
src="https://github.com/user-attachments/assets/650c5e4a-8899-4073-95d2-91ba21b94e8b"
/>