feat: Connect websocket in background
This change removes the websocket connection logic from the UI and moves it to the Rust backend.

The websocket connection is now initiated in the background when the app starts up.
The WebsocketEventsManager now handles the websocket connection and reconnection.
This improves the user experience by removing the need for the UI to manage the websocket connection.
Removes un-needed `initialiseSocket` calls and related code.