fix: fix missing telemetry messages that happen before service starts (#2820)
I noticed an `if let Some(x) =` check around the tx_channel when trying
to send a telemetry event. This lead me to refactor it a bit so that the
tx_channel was always populated.

I changed the looping logic around telemetry as well as removing
retries, which seem unnecessary, since we will send again later.