fix(ui): paper wallet x pin input overlap (#2816)
Description
---

- used updated `Dialog` component wrappers for `GreenModal` and removed
the custom cover and styling (incl. the problematic z-index)
- adjusted `GreenModal` styling to suit
- added new `wrapper` variant yo `Dialog` - so you can add any super
customized modal components in there and still be within the correct
floating node heirarchy

Motivation and Context
---

- fixes https://github.com/tari-project/universe/issues/2814

How Has This Been Tested?
---
- locally


**smallest screen size:**


https://github.com/user-attachments/assets/ed328997-d210-45ad-be16-12af9630dcb3

**all screen sizes:**


https://github.com/user-attachments/assets/ee32cc1e-0ab5-447b-aa31-f06bbddd2a51



What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->