fix(ui): make a root avatar component and integrate with crew rewards (#2815)
This PR moves the Avatar component from the Airdrop Sidebar to be a
global component.

Its now being used for all avatars in the Crew Rewards widget as well as
the sidebar. It requires a username to be passed in order to generate an
avatar for the user when the image field is blank.

<img width="1292" height="1406" alt="CleanShot 2025-09-04 at 17 32
25@2x"
src="https://github.com/user-attachments/assets/e0e4420f-f949-4159-966b-6d6755d74f5f"
/>