refactor: Improve keep-alive logging
Reduced verbosity of keep-alive logging. The log message now only indicates that a keep-alive message is being sent, rather than including the message contents. The detailed message content was deemed unnecessary for regular operation and clutters the logs. The confirmation tick message was also removed as it added little value.